+++ tests/testthat/test-translate-sql-string.R
@@ -17,14 +17,24 @@ test_that("sql_str_sub works as expected", {
 
   expect_equal(str_sub(x), sql("SUBSTR(`x`, 1)"))
   expect_equal(str_sub(x, 1), sql("SUBSTR(`x`, 1)"))
-  expect_equal(str_sub(x, -1), sql("SUBSTR(`x`, -1)"))
+  expect_equal(str_sub(x, -1), sql("SUBSTR(`x`, LENGTH(`x`))"))
   expect_equal(str_sub(x, 2, 4), sql("SUBSTR(`x`, 2, 3)"))
   expect_equal(str_sub(x, 2, 2), sql("SUBSTR(`x`, 2, 1)"))
   expect_equal(str_sub(x, 2, 0), sql("SUBSTR(`x`, 2, 0)"))
   expect_equal(str_sub(x, 1, -2), sql("SUBSTR(`x`, 1, LENGTH(`x`) - 1)"))
   expect_equal(str_sub(x, 3, -3), sql("SUBSTR(`x`, 3, LENGTH(`x`) - 4)"))
-  expect_equal(str_sub(x, -3, 0), sql("SUBSTR(`x`, -3, 0)"))
-  expect_equal(str_sub(x, -3, -3), sql("SUBSTR(`x`, -3, 1)"))
+  expect_equal(str_sub(x, -3, 0), sql("SUBSTR(`x`, LENGTH(`x`) - 2, 0)"))
+  expect_equal(str_sub(x, -3, -3), sql("SUBSTR(`x`, LENGTH(`x`) - 2, 1)"))
+})
+
+test_that("sql_str_sub can require length paramter", {
+  local_con(simulate_dbi())
+  x <- ident("x")
+  str_sub <- sql_str_sub("SUBSTR", optional_length = FALSE)
+
+  expect_equal(str_sub(x), sql("SUBSTR(`x`, 1, LENGTH(`x`))"))
+  expect_equal(str_sub(x, 1), sql("SUBSTR(`x`, 1, LENGTH(`x`))"))
+  expect_equal(str_sub(x, -1), sql("SUBSTR(`x`, LENGTH(`x`), 1)"))
 })
 
 test_that("str_sub() returns consistent results", {